requirements.txt
spotapi
pymongo
redis
websockets
python-dotenv
pymongo[srv]==3.12
flask